+++ MatchEngineHockey/Goal_hockey.h
@@ -8,14 +8,14 @@
 {
     Game_hockey *GameController;
 	SPImage *myImage;
-    int team;
-    int side;
+    NSInteger team;
+    NSInteger side;
 }
 @property (nonatomic, strong) Game_hockey *GameController;
-@property (nonatomic) int team;
-@property (nonatomic) int side;
-- (id)init:(int)s;
-+ (Goal_hockey *)initGoal:(int)s;
+@property (nonatomic) NSInteger team;
+@property (nonatomic) NSInteger side;
+- (id)init:(NSInteger)s;
++ (Goal_hockey *)initGoal:(NSInteger)s;
 - (BOOL)checkGoal:(Ball_hockey *)ball;
 - (Player_hockey *)getClosestPlayer;
 - (SPPoint*)point;
